requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
appnope==0.1.4
asttokens==3.0.0
attrs==25.3.0
certifi==2025.8.3
charset-normalizer==3.4.3
comm==0.2.3
contourpy==1.3.3
cycler==0.12.1
dateparser==1.2.2
debugpy==1.8.17
decorator==5.2.1
et_xmlfile==2.0.0
executing==2.2.1
fonttools==4.60.0
frozenlist==1.7.0
idna==3.10
ipykernel==6.30.1
ipython==9.5.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
jupyter_client==8.6.3
jupyter_core==5.8.1
kiwisolver==1.4.9
matplotlib==3.10.6
matplotlib-inline==0.1.7
multidict==6.6.4
nest-asyncio==1.6.0
numpy==2.3.3
openpyxl==3.1.5
packaging==25.0
pandas==2.3.2
parso==0.8.5
pexpect==4.9.0
pillow==11.3.0
platformdirs==4.4.0
prompt_toolkit==3.0.52
propcache==0.3.2
psutil==7.1.0
ptyprocess==0.7.0
pure_eval==0.2.3
pycryptodome==3.23.0
Pygments==2.19.2
pyparsing==3.2.5
python-binance==1.0.29
python-dateutil==2.9.0.post0
pytz==2025.2
pyzmq==27.1.0
regex==2025.9.18
requests==2.32.5
six==1.17.0
stack-data==0.6.3
tornado==6.5.2
traitlets==5.14.3
typing_extensions==4.15.0
tzdata==2025.2
tzlocal==5.3.1
urllib3==2.5.0
wcwidth==0.2.14
websockets==15.0.1
yarl==1.20.1
